Réorganisation de la couche API pour authentifier l'utilisateur et gérer les droits

+++ src/store/index.js
@@ -1,7 +1,6 @@
 import Vue from 'vue'
 import Vuex from 'vuex'
 import createLogger from 'vuex/dist/logger'
-import AuthenticationAPI from '@/api/authentication'
 import actions from '@/store/actions'
 import getters from '@/store/getters'
 import mutations from '@/store/mutations'
@@ -10,14 +9,20 @@ import { createInitialStoreState } from '@/store/state'
 
 Vue.use(Vuex)
 
-export async function createStore () {
+export async function createStore (options = {}) {
+  const api = options.api
+  if (!api) {
+    throw new Error('Missing `api` option')
+  }
+
   const sessionToken = sessionStorage.load()
-  const authenticationInfos = await AuthenticationAPI.authenticate(sessionToken)
+
+  const authenticationInfos = await api.authentication.authenticate(sessionToken)
   if (!authenticationInfos.valid) {
     sessionStorage.delete()
   }
 
-  return new Vuex.Store({
+  const store = new Vuex.Store({
     state: createInitialStoreState(authenticationInfos),
     getters,
     mutations,
@@ -25,4 +30,12 @@ export async function createStore () {
     strict: process.env.NODE_ENV !== 'production',
     plugins: [createLogger()]
   })
+
+  // expose the api in the store
+  store.$api = api
+
+  // expose the store in the api
+  api.store = store
+
+  return store
 }
